updated prisma migrate usage in docs (#419)
