+++ app/core/navs/documentation.json
@@ -53,6 +53,7 @@
       "head-component",
       "document-component",
       "preview-mode",
+      "static-html-export",
       "code-splitting",
       "get-static-props",
       "get-static-paths",
@@ -72,7 +73,8 @@
       "link",
       "route-params-query",
       "router",
-      "api-routes"
+      "api-routes",
+      "shallow-routing"
     ]
   },
   {
@@ -167,6 +169,7 @@
       "cli-dev",
       "cli-start",
       "cli-build",
+      "cli-export",
       "cli-db",
       "cli-prisma",
       "cli-generate",
