+++ app/pages/docs/database-overview.mdx
@@ -40,8 +40,8 @@ model Task {
 > If you need,
 > [reference the Prisma Schema documentation](https://www.prisma.io/docs/reference/tools-and-interfaces/prisma-schema/data-model)
 
-2. Then run `blitz prisma migrate dev --preview-feature` in your terminal
-   to apply the changes
+2. Then run `blitz prisma migrate dev` in your terminal to apply the
+   changes
 3. Now you can import `db` from `db/index.ts` and create a model like
    this:
    - `db.project.create({data: {name: 'Hello'}})`
@@ -67,6 +67,5 @@ datasource db {
    your setup, you may need to modify the URL.
 3. Update `.env.test.local` as well to use PostgreSQL in your test runs.
 4. Delete the `db/migrations` folder
-5. Run `blitz prisma migrate dev --preview-feature`. This command will
-   create the database (if it does not already exist) and tables based on
-   your schema.
+5. Run `blitz prisma migrate dev`. This command will create the database
+   (if it does not already exist) and tables based on your schema.
