+++ app/pages/docs/deploy-heroku.mdx
@@ -62,7 +62,7 @@ heroku git:remote --app <APP_NAME>
    following content.
 
 ```
-release: npx blitz prisma migrate deploy --preview-feature
+release: npx blitz prisma migrate deploy
 web: npm run start:production
 ```
