+++ app/pages/docs/deploy-render.mdx
@@ -4,10 +4,10 @@ sidebar_label: To Render.com
 ---
 
 1. Add one of the `render.yaml` files shown below
-2. Push code to your github repo
+2. Push code to your GitHub repo
 3. Log in to [Render.com](https://render.com)
 4. Click on the "YAML" menu item, then click the "New from YAML" button
-5. Connect your github account then select your blitz app repo
+5. Connect your GitHub account then select your blitz app repo
 6. Click approve
 7. Go to "Services", click on your new Blitz service, click on
    "Environment"
@@ -49,7 +49,7 @@ services:
     buildCommand:
       yarn --frozen-lockfile --prod=false &&
       blitz build &&
-      blitz prisma migrate deploy --preview-feature
+      blitz prisma migrate deploy
     startCommand: blitz start
     envVars:
       - key: NODE_ENV
