+++ app/pages/docs/error-handling.mdx
@@ -26,7 +26,8 @@ application.
   - Default `message`: "This could not be found"
 
 To use, import from `blitz` and use like any JavaScript Error. If you're
-curious, you can [see the source code for these.](TODO)
+curious, you can
+[see the source code for these](https://github.com/blitz-js/blitz/blob/canary/packages/core/src/errors.ts).
 
 ```ts
 import {AuthenticationError} from "blitz"
@@ -164,5 +165,4 @@ can handle the error in your form submit handler like this.
 
 ## Error Pages {#error-pages}
 
-See the [Error Pages documentation](./error-pages) to learn about custom
-404 and 500 error pages.
+See the [Error Pages documentation](./error-pages) to learn about custom error pages and error fallback components.
