+++ app/pages/docs/error-pages.mdx
@@ -56,6 +56,72 @@ export default Error
 > get an error with the call stack to know where the error originated
 > from.
 
+## Customize the Error Boundary Fallback Component {#custom-error-boundary-component}
+
+The default `app/pages/_app.tsx` configures an error boundary which catches errors that happen during render, on the client side. This will also catch errors within `useQuery` or `useMutation` because they happen in suspense.
+
+In essence, the default configuration is:
+
+```tsx
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <ErrorBoundary
+      FallbackComponent={RootErrorFallback}
+      {...etc}
+    >
+      {getLayout(<Component {...pageProps} />)}
+    </ErrorBoundary>
+  )
+}
+```
+
+The `RootErrorFallback` is rendered when a error happens within render, or a suspended mutation or query. By default:
+
+```tsx
+function RootErrorFallback({ error, resetErrorBoundary }: ErrorFallbackProps) {
+  if (error instanceof AuthenticationError) {
+    return <LoginForm onSuccess={resetErrorBoundary} />
+  } else if (error instanceof AuthorizationError) {
+    return (
+      <ErrorComponent
+        statusCode={error.statusCode}
+        title="Sorry, you are not authorized to access this"
+      />
+    )
+  } else {
+    return (
+      <ErrorComponent statusCode={error.statusCode || 400} title={error.message || error.name} />
+    )
+  }
+}
+```
+
+### Authentication Error Component (401) {#custom-authentication-error-component}
+
+If your query, mutation, or render function `throw new AuthenticationError()` when the user is not authenticated, you can detect that within the error fallback:
+
+```tsx
+function RootErrorFallback({ error, resetErrorBoundary }: ErrorFallbackProps) {
+  if (error instanceof AuthenticationError) {
+    return <MyCustomError error={error} onRetry={resetErrorBoundary} />
+  }
+  return <ErrorComponent statusCode={error.statusCode || 400} title={error.message || error.name} />
+}
+```
+
+### Authorization Error Component (403) {#custom-authorization-error-component}
+
+Your query, mutation, or render function may call `throw new AuthorizationError()` when a user is authenticated but not authorized to do something. The error fallback may handle that:
+
+```tsx
+function RootErrorFallback({ error, resetErrorBoundary }: ErrorFallbackProps) {
+  if (error instanceof AuthorizationError) {
+    return <MyCustomError error={error} onRetry={resetErrorBoundary} />
+  }
+  return <ErrorComponent statusCode={error.statusCode || 400} title={error.message || error.name} />
+}
+```
+
 ## Reusing the built-in error page {#reusing-the-built-in-error-page}
 
 You can render the built-in error page by importing `ErrorComponent`.
