+++ app/pages/docs/how-the-community-operates.mdx
@@ -57,7 +57,7 @@ smoothing rough edges and fixing bugs. However, anyone else is more than
 welcome to continue adding major features!
 
 Our [GitHub issues](https://github.com/blitz-js/blitz/issues) are where
-you can find our roadmap. Many planned features have github issues, but
+you can find our roadmap. Many planned features have GitHub issues, but
 not necessarily all of them do.
 
 ## How to Propose Features/Changes {#how-to-propose-features-changes}
