+++ app/pages/docs/link.mdx
@@ -48,8 +48,8 @@ export default Home
   `false`
 - [`scroll`](#disable-scrolling-to-the-top-of-the-page) - Scroll to the
   top of the page after a navigation. Defaults to `true`
-- [`shallow`](#todo) - Update the path of the current page without
-  rerunning [`getStaticProps`](./get-static-props) or
+- [`shallow`](./shallow-routing) - Update the path of the current page
+  without rerunning [`getStaticProps`](./get-static-props) or
   [`getServerSideProps`](./get-server-side-props). Defaults to `false`
 
 External URLs, and any links that don't require a route navigation using
