+++ app/pages/docs/postgres.mdx
@@ -80,5 +80,5 @@ Given these values your `DATABASE_URL` should look like this
 }
 ```
 
-4. Run `blitz prisma migrate dev --preview-feature` to get your new
-   database to the latest version of your migrations
+4. Run `blitz prisma migrate dev` to get your new database to the latest
+   version of your migrations
