+++ package.json
@@ -2,14 +2,13 @@
   "name": "blitz-site-v2",
   "version": "1.0.0",
   "scripts": {
-    "dev": "yarn gen-nav && blitz dev",
+    "dev": "blitz dev",
     "start": "blitz start",
-    "build": "yarn gen-nav && blitz build",
+    "build": "blitz build",
     "lint": "yarn lint:code && yarn lint:docs",
     "lint:code": "eslint --ignore-path .gitignore --ext .js,.ts,.tsx .",
     "lint:docs": "alex app/pages/docs/**/*.mdx",
     "english-slugify": "node scripts/english-slugify.js",
-    "gen-nav": "node scripts/generate-documentation-nav.js && prettier -w app/core/navs/*",
     "postinstall": "husky install"
   },
   "lint-staged": {
@@ -51,6 +50,7 @@
     "@mdx-js/loader": "^1.6.22",
     "@next/bundle-analyzer": "^10.0.8",
     "@svgr/webpack": "^5.4.0",
+    "@tailwindcss/jit": "0.1.3",
     "@tailwindcss/typography": "0.4.0",
     "alex": "9.1.0",
     "autoprefixer": "^10.2.5",
